Add support for all national clouds:
- Global: graph.microsoft.com
- Germany: graph.microsoft.de
- China: microsoftgraph.chinacloudapi.cn
- US Government L4: graph.microsoft.us
- US Government L5 (DOD): dod-graph.microsoft.us

Note that this is a breaking change from v0.1.0 as the signatures for
all the clients have changed. If you are using the global cloud, you do
not need to specify this when creating a new client as it is the default.
However, you do need to specify a cloud environment when acquiring an
access token using auth.NewAuthorizer.