Refactor clients and models into a single package
This makes for a better experience when importing, for example the
`models` package used a very generic name and could easily conflict
with a models package local to a project. By flattening clients and
models into a single package, it also makes for easier inspection in
your code, since both now reside in the `msgraph` package. And whilst
we're at it, might as well consolidate the `base` package too, since
there's no significant benefit to maintaining a separate package for
that.

The AAD Graph client has simialr treatment, although there are no
endpoint-specific clients or models for that [yet].

The `hamilton/base/odata` package has been moved to `hamilton/odata`.

This is a breaking change and all users will need to redo their imports,
but on the bright side it's now just a single import which is unlikely
to change again in the near future.