Fix for handling OData `value` field having a type that is not an array
Previously we assumed that the OData `value` field was always a json
array, which raised an error in the msgraph.Client{}.Get() method
during paging, as we would try to merge `value`s from different pages
and present as a single result.

The odata.OData{} struct now handles this field with an empty interface,
and the msgraph.Client{}.Get() method tries to assert this as holding a
slice, then merging the `value`s from multiple pages if this assertion
holds true for every page.

This is technically a breaking change on odata.OData{}.Value, but since
this was previously typed as `*[]json.RawMessage` it would have to be
parsed with a custom struct anyway. This new implementation is slower
however, sometimes noticably with large resultsets.